CoreTests: Updates to LNS records don't update register_height
Register height is only updated when a mapping is renewed or when first
purchasing the mapping, i.e the register height. Before we assumed,
register_height was changed on every update, but, you are given the TXID
of the transaction already which by extension gives you the
register_height so that was also somewhat redundant.