Store base64 hashes into the LNS DB
Everywhere still uses crypto::hash for the name_hash due to being the
most compact storage form except for the LNS DB which uses base64 so
that it can be indexes in SQL for speed.

But at the boundary between requesting and accessing the DB entry, the
hash gets converted into a base64 representation of the hash.

We opt for base64 as it is more compact than storing the hex
representation.